Checks whether a log file exists and is readable
Checks whether the pages and media changelog file exits and is
readable before access to and read in the lines.

Especially for the media changelog file it is a common case that
they doesn't exist after a fresh Dokuwiki installation.

Error was:

Invalid argument supplied for foreach() in .../syntax.php on line 207
array_merge(): Argument #2 is not an array in .../syntax.php on line 210

Signed-off-by: Stephan Linz <linz@li-pro.net>